backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
livekit-agents==1.0.0
livekit-plugins-google==1.0.0
livekit==1.0.0
python-dotenv==1.0.1
httpx==0.27.0
fhir.resources==7.1.0
supabase==2.10.0
numpy>=1.26.0
pandas==2.2.3
scikit-learn==1.5.2
pydantic==2.9.0
python-multipart==0.0.12
